// src/modules/invoice/usecase/generate/generate-invoice.usecase.dto.ts
type ItemInput = {
  id: string;
  name: string;
  price: number;
};

export interface GenerateInvoiceUseCaseInputDto {
  name: string;
  document: string;
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  items: ItemInput[];
}

type ItemOutput = {
  id: string;
  name: string;
  price: number;
};

export interface GenerateInvoiceUseCaseOutputDto {
  id: string;
  name: string;
  document: string;
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  items: ItemOutput[];
  total: number;
}